Hide "English support only" notice for forum posts
This hides the "Note: Support is only available in English at the moment" notice on /help/contact if the user is only eligible for forum support.

Because forum posts are directed to the relevant locale forum (for popular languages), we shouldn't tell users that support is only available in English.